Complete rewrite
Something bad happened to the project file, preventing building. With
less than a week left before the competition, and freshly installing
Eclipse not fixing the problem, I copied all of the code into a new
blank file. Jetson and GRIP code was not included. I'm honestly not sure
what's going to happen when I commit and push this.